season1/requirements.txt
datasets==2.0.0
deepspeed==0.6.0
fairscale==0.4.6
filelock==3.6.0
numpy
packaging==21.3
scikit_learn==1.0.2
tqdm
transformers==4.17.0
sacrebleu==1.4.14
faiss-gpu==1.7.2
pytorch_lightning==1.6.0